try another fix

+++ src/android/utility/base/strings.cpp
@@ -21,10 +21,10 @@
 #include<cstring>
 std::vector<std::string> android::base::Split(const std::string& s, const std::string& delimiters) {
 	std::vector<std::string>result;
-	std::string str(s);
-	std::string  token(delimiters);
-	while (str.size()) {
-		unsigned int index = str.find(token);
+ 	std::string str(s);
+ 	std::string  token(delimiters);
+ 	while (str.size()) {
+		size_t index = str.find(token);
 		if (index != std::string::npos) {
 			result.push_back(str.substr(0, index));
 			str = str.substr(index + token.size());
